+++ CHANGELOG.md
@@ -23,3 +23,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -core/interfaces.go - GetGiesCourses added to Services interface
 -core/interfaces.go - Defined GiesCourses interface
 -core/services.go - added getGiesCourses implementation to application
+
+## [1.2.0] - 8/4/2022
+### Fixed
+### Added
+-Endpoint courses/studentcourses was added to return classes and their locations for students for selected semester
+-driven/courses/uiuc_courses.go
+-Endpont /termsessions/listcurrent was added to return a list of currently selectable term sessions
+-driven/terms/uiuc_termsessions.go
+-driver/rest/termsessionapi.go
+-core/model/termsessions.go
+
+
+### Changed
+-core/interfaces.go - GetStudentCourses added to Services interface
+-core/interfaces.go - Defined student courses interface
+-core/services.go - added getStudentCourses implementation to application
+-core/interfaces.go - GetTermSessions added to Services interface
+-core/interfaces.go - Defined TermSessions interface
+-core/services.go - added getTermSessions implementation to application
+-driver/rest/coursesapi.go
+-core/model/courses.go
+-drive/adapter.go - added routes for /courses/studentcourses and termsessions/listcurrent
+-main.go - added term session adapter to application
+-application.go - added a term session adapter
+-services.go - implemented getStudentCourses and getTermSessions